Rename main to streams, streams to streamtypes
